Update Nbody example according to the new code in SRA
Issue #96

Reviewed by: Diego Pinheiro.

The kernel was modified by Steven Eliuk (see issue #260 in SRA/webkit-webcl repo),
because it was running slower than serial JS when CPU was used. So, this new code
is being imported here and the device can be changed using the radio buttons.

When CPU is selected, the CL-GL interoperability is disabled. This change increases
frame rate.