Disabling CL<->GL interop when using CPU device. Issue #97
Reviewed by: Diego Pinheiro

When CPU device is selected the interop button will be
hidden. To look better on the web page the interop button
position was changed to be on the side of GPU device. Also,
removed the alert when a error occurs in draw because it was
very annoying and impossible to stop the pop up windows.